+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Add support for renaming multiple columns at once in the same table
 - Fix quoting table/column names across the library
 - Fix deffered foreign keys support in `add_foreign_key` (Active Record >= 7)
 - Reset attempts of failing jobs before executing background migration inline